Load summary stats from raw keyspace

+++ src/main/scala/at/ac/ait/Transformation.scala
@@ -370,7 +370,7 @@ class Transformation(spark: SparkSession) {
 
   def summaryStatistics(
       lastBlockTimestamp: Int,
-      noBlocks: Long,
+      noBlocks: Int,
       noTransactions: Long,
       noAddresses: Long,
       noAddressRelations: Long,
